Changed future into adjacent possible environment
Based on comment by Dave Snowden: Thanks and it is interesting - from a CAS perspective I think we would focus on mapping the adjacent possible rather than setting visions for the future and would use a mass generated fitness landscape - but the principe applies.  I like the sketch do you plan to publish it?